feat(CEPA-158): Create course detail - Add getCourse, getCourseDetail - Create CourseVideoResponse, CourseDetailResponse - Create CourseVideoMapper

+++ src/main/java/com/example/courseservice/data/dto/request/CourseRequest.java
@@ -13,5 +13,5 @@ public class CourseRequest {
     private String name;
     private Double price;
     private String subject;
-    private int levelId;
+    private long levelId;
 }
